refactor: update package declaration for asset paths
